record create modified and project seeder 2024/18/4

+++ database/migrations/2024_03_21_115848_create_budgets_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('timeline_id')->constrained('timelines', 'id')->onDelete('cascade')->onUpdate('cascade');
             $table->string('name');
-            $table->bigInteger('amount');
+            $table->decimal('amount',60,4);
  //todo problem
             $table->timestamps();
         });
